промпты в gemini, хэштеги для видео

+++ backend/video/views.py
@@ -3,8 +3,9 @@ from drf_spectacular.utils import extend_schema
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from .serializers import VideoRequestSerializer
-from .highlights import analyze_video_with_scores
+from .highlights import analyze_video_with_scores, get_random_frame
 from .utils import convert_to_seconds, cut_video_segments, create_zip_from_files, send_zip_via_email
+import uuid
 
 @extend_schema(
     summary="Загрузка видео и email для анализа",
@@ -19,26 +20,38 @@ def upload_video(request):
     
     video = serializer.validated_data["video"]
     email = serializer.validated_data["email"]
-    path = video.temporary_file_path()
-
-    result = analyze_video_with_scores(path)
+    video_path = video.temporary_file_path()
     
+    result = analyze_video_with_scores(video_path)
     highlights = result["highlights"]
     timestamps = [convert_to_seconds(h["time"]) for h in highlights]
     
-    segment_paths = cut_video_segments(path, timestamps)
+    segment_paths = cut_video_segments(video_path, timestamps)
     zip_path = "/tmp/highlights.zip"
-    
     create_zip_from_files(segment_paths, zip_path)
+    
+    img_path = None
     try:
-        send_zip_via_email(email, zip_path)
+        image = get_random_frame(video_path)
+        if image:
+            img_path = f"/tmp/frame_{uuid.uuid4().hex}.jpg"
+            image.save(img_path)
+
+            send_zip_via_email(email, zip_path, img_path)
+        else:
+            raise Exception("Не удалось получить кадр из видео")
     except Exception as e:
         return Response({"status": "error", "message": f"Ошибка при отправке письма: {e}"}, status=500)
-
-    os.remove(path)
-    os.remove(zip_path)
-    for p in segment_paths:
-        os.remove(p)
+    finally:
+        if os.path.exists(video_path):
+            os.remove(video_path)
+        if os.path.exists(zip_path):
+            os.remove(zip_path)
+        if img_path and os.path.exists(img_path):
+            os.remove(img_path)
+        for p in segment_paths:
+            if os.path.exists(p):
+                os.remove(p)
 
     return Response({"status": "ok", "message": "Анализ завершён, нарезки отправлены на email."})
